kernel: generic: 5.15: allow to choose QCOM_QMI_HELPERS
QMI helpers are required by ath11k, but the symbol is not user selectable
as usually ath11k itself will just select the symbol.
This doesnt work in our case as backports cannot do the same, you can
only depend on a symbol.

So to allow building in OpenWrt, lets make the symbol selectable again
so it can be packaged as a kmod and added as a dependency to ath11k.

Disable the symbol in the generic config so that builds dont get broken
by the build system asking whether to enable the symbol on each build.

Signed-off-by: Janusz Dziedzic <janusz.dziedzic@gmail.com>
Signed-off-by: Robert Marko <robert.marko@sartura.hr>